feat(app-lists-db): PRD-245 US-06 relocate lists-owned schemas from db-types (audit H6/H7)
Smallest US in PRD-245 (db-types decomposition / audit H6+H7). Relocates
the 2 lists-owned drizzle tables (`lists`, `list_items`) out of
`packages/db-types/src/schema/` into `packages/app-lists-db/src/schema/`,
alongside the `lists-row-schemas.ts` sibling. `@pops/db-types` becomes
the consumer; `@pops/app-lists-db` becomes the canonical owner.

No cross-pillar FK references to drop — the lists tables are self-contained.
Mirrors the proven pattern from #3264 (US-04 media).

Cycle-break order (post-#3256 lesson):
- Drop `@pops/db-types` from `packages/app-lists-db/package.json`. After
  the schema move app-lists-db has zero remaining references.
- Add `@pops/app-lists-db` to `packages/db-types/package.json`. db-types'
  `src/schema/index.ts` re-exports the relocated tables as a transition
  shim; `src/lists.ts` continues to host the `InferSelectModel<typeof …>`
  aliases (`ListRow`, `ListItemRow`, `ListInsert`, `ListItemInsert`,
  `ListKind`, `ListItemRefKind`) so the public surface of `@pops/db-types`
  stays identical until US-08.
- Drop `lists-row-schemas.ts` from `packages/db-types/src/schema/row-schemas.ts`.
- Reorder every per-app Dockerfile that builds db-types so app-lists-db
  builds first. Affected: pops-api, pops-shell, pops-mcp (already
  copied app-lists-db source — just reorder); pops-cerebrum-api,
  pops-core-api, pops-finance-api, pops-food-api, pops-inventory-api,
  pops-lists-api, pops-media-api (added COPY for package.json + src +
  tsconfig and inserted the build step before db-types).

Cycle check: `pnpm install` reports zero new cyclic workspace deps.

Consumer repoint: zero source files needed flipping — the lists module
under `apps/pops-api/src/modules/lists/` already imports
`@pops/app-lists-db` and `@pops/lists-db` directly (no `@pops/db-types`
references to repoint).

Smoke test: `packages/app-lists-db/src/__tests__/schema-smoke.test.ts`
asserts both relocated tables resolve from `@pops/app-lists-db` with
their expected drizzle SQL `name`.

Verification:
- pnpm --filter @pops/app-lists-db build: clean
- pnpm --filter @pops/app-lists-db test: 2 files / 43 tests pass
  (includes new schema-smoke.test.ts)
- pnpm --filter @pops/db-types build: clean
- pnpm --filter @pops/db-types lint:row-schemas: 48 sibling files / 68
  tables check out
- pnpm --filter @pops/api build: clean
- pnpm --filter @pops/api test: 312 files / 4899 tests pass
- pnpm typecheck (root): 73/73 turbo tasks pass
- pnpm build (root): 37/37 turbo tasks pass
- pnpm lint: warnings only (pre-existing)
- pnpm format:check: clean

References:
- PRD: docs/themes/13-pillar-finale/prds/245-db-types-decomposition/README.md
- US: docs/themes/13-pillar-finale/prds/245-db-types-decomposition/us-06-relocate-lists-schemas.md
- Reference PR: #3264 (US-04 media)
