+++ apps/pops-mcp/Dockerfile
@@ -87,10 +87,12 @@ COPY apps/pops-mcp/src ./apps/pops-mcp/src
 
 # Build shared packages first (module-registry needs db-types + types;
 # app-food-db is consumed by pops-api routers).
-# media-db owns the canonical media table schemas and is now a leaf
-# package; db-types re-exports them via a transition shim, so media-db
-# must build BEFORE db-types (PRD-245 US-04 / audit H6).
+# media-db / app-lists-db own canonical schemas; db-types re-exports them
+# via a transition shim, so both MUST build BEFORE db-types
+# (PRD-245 US-04 media + US-06 lists / audit H6).
 WORKDIR /app/packages/media-db
+RUN pnpm build
+WORKDIR /app/packages/app-lists-db
 RUN pnpm build
 WORKDIR /app/packages/db-types
 RUN pnpm build
@@ -111,8 +113,6 @@ RUN pnpm build
 WORKDIR /app/packages/food-db
 RUN pnpm build
 WORKDIR /app/packages/lists-db
-RUN pnpm build
-WORKDIR /app/packages/app-lists-db
 RUN pnpm build
 WORKDIR /app/packages/app-food-db
 RUN pnpm build
