+++ packages/app-lists-db/package.json
@@ -19,8 +19,9 @@
     "test:watch": "vitest"
   },
   "dependencies": {
-    "@pops/db-types": "workspace:*",
-    "drizzle-orm": "^0.45.2"
+    "drizzle-orm": "^0.45.2",
+    "drizzle-zod": "^0.8.3",
+    "zod": "^4.4.3"
   },
   "devDependencies": {
     "@types/better-sqlite3": "^7.6.12",
