+++ packages/db-types/package.json
@@ -28,6 +28,7 @@
     "test": "echo \"(no tests in @pops/db-types)\""
   },
   "dependencies": {
+    "@pops/app-lists-db": "workspace:*",
     "@pops/media-db": "workspace:*",
     "drizzle-orm": "^0.45.2",
     "drizzle-zod": "^0.8.3",
