+++ packages/db-types/src/schema/index.ts
@@ -65,7 +65,7 @@ export {
 } from './food.js';
 export { substitutions } from './food-substitutions.js';
 export { homeInventory } from './inventory.js';
-export { listItems, lists } from './lists.js';
+export { listItems, lists } from '@pops/app-lists-db';
 export { itemConnections } from './item-connections.js';
 export { itemFixtureConnections } from './item-fixture-connections.js';
 export { itemDocuments } from './item-documents.js';
